+++ lib/Utils/Engines/EngineInterface.php
@@ -105,7 +105,7 @@ interface Engines_EngineInterface {
      * @param $target
      * @param $sentence
      * @param $translation
-     * @return int|null
+     * @return float|null
      */
-    public function getQualityEstimation($source, $target, $sentence, $translation) : ?int;
+    public function getQualityEstimation($source, $target, $sentence, $translation) : ?float;
 }
